typeset_numbers {ggpmisc}	R Documentation

Typeset/format numbers preserving trailing zeros

Description

Typeset/format numbers preserving trailing zeros

Usage

typeset_numbers(eq.char, output.type)

Arguments

eq.char
character A polynomial model equation as a character string.

output.type
character One of "expression", "latex", "tex", "text", "tikz", "markdown".

Value

A character string.

Note

exponential number notation to typeset equivalent: Protecting trailing zeros in negative numbers is more involved than I would like. Not only we need to enclose numbers in quotations marks but we also need to replace dashes with the minus character. I am not sure we can do the replacement portably, but that recent R supports UTF gives some hope.

[Package ggpmisc version 0.6.3 Index]